Fixed runtime issues after refactoring
